EES-5446 - corrected long-running test orchestration.
