+++ infrastructure/templates/public-api/types.bicep
@@ -41,11 +41,16 @@ type ResourceNames = {
 }
 
 @export()
+type IpRange = {
+  name: string
+  cidr: string
+}
+@export()
 type FirewallRule = {
   name: string
   cidr: string
-  priority: int?
-  tag: string?
+  priority: int
+  tag: string
 }
 
 @export()
